Documented continue_training and moved it into serialization
