--- bin/blocks-continue
@@ -1,15 +1,0 @@
-#!/usr/bin/env python
-import logging
-from argparse import ArgumentParser
-
-from blocks.scripts import continue_training
-
-if __name__ == "__main__":
-    logging.basicConfig()
-
-    parser = ArgumentParser("Continues your pickled main loop")
-    parser.add_argument(
-        "path", help="A path to a file with a pickled main loop")
-    args = parser.parse_args()
-
-    continue_training(args.path)
