--- bin/blocks-dump
@@ -1,23 +1,0 @@
-#!/usr/bin/env python
-import logging
-from argparse import ArgumentParser
-
-from blocks.scripts import dump
-
-if __name__ == "__main__":
-    logging.basicConfig()
-
-    parser = ArgumentParser(
-        "Dumps your pickled main loop. Dumps are a cross-platform way of"
-        " sharing results of your experiments orthogonal to the standard"
-        " seriaziation approch. For more information, see documentation"
-        " of blocks/dump.py module.")
-    parser.add_argument(
-        "pickle_path", help="A path to a file with a pickled main loop")
-    parser.add_argument(
-        "dump_path", nargs="?",
-        help="A path where you want to dump."
-             " When not given, the root of 'pickle_path' is used.")
-    args = parser.parse_args()
-
-    dump(args.pickle_path, args.dump_path)
